pyproject.toml
[project]
name = "faceless-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.10.5",
    "asyncpg>=0.29.0",
    "bcrypt>=4.2.0",
    "boto3>=1.35.19",
    "captacity>=0.3.1",
    "fastapi>=0.114.2",
    "greenlet>=3.1.0",
    "moviepy>=1.0.3",
    "openai>=1.45.0",
    "opencv-python>=4.10.0.84",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.9",
    "pydantic-settings>=2.5.2",
    "pydantic>=2.9.1",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.9",
    "replicate>=0.33.0",
    "ruff>=0.6.5",
    "socksio>=1.0.0",
    "sqlalchemy>=2.0.34",
    "supabase>=2.7.4",
    "torch>=2.3.1",
    "torchaudio>=2.3.1",
    "uvicorn>=0.30.6",
]
